update $streamName.tsx to read new data properties from updates
we are loading streams a bit differently now.

triggering tweet and follower network updates automatically.

tweet update times are stored on the neo4j relationship between a stream and a user, so we know what range of dates for tweets we have on a per user basis. follows is still done on a stream level, but this could be hard to convey to the users, so may need to replicate that for follows as well