new getStreamTweets function pulling from listTweets endpoint
much simpler logic.
i still need to index the ranges from which we have pulled tweets from a list, and which members were part of the list

Not sure how I will do this quite yet. will not be as precise as managing tweets on a per user basis, but this will still simplify that code and improve performance